pyproject.toml
[project]
name = "lmg-ai-agents-examples"
version = "0.1"
requires-python = ">= 3.13,<3.14"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
agent-framework = "^1.0.0b251105"
azure-keyvault-secrets = "^4.10.0"
